Add a case for using play with docker
Play with docker doesn't give write access for creating the subdirectories in a volume to copy the tutorial to, so just put it back where it was since the files are all on the server anyway.  To indicate that you need to add play when running it to bypass the volume.