Use the actr7.x/next version now
Changed to using the actr7.x/next/actr7.container.zip, but really need to adjust how it gets ACT-R in general since the .zip gets cached in a docker image and it doesn't download a new one after the first build.  Touching the Dockerfile every time isn't the "right" way to handle this.